[mmu] Buffer: support large sets/ways values

+++ src/main/java/plugins/mmu/ru/ispras/microtesk/mmu/translator/ir/spec/builder/MmuSpecBuilder.java
@@ -143,8 +143,8 @@ public final class MmuSpecBuilder implements TranslatorHandler<Ir> {
 
       final MmuDevice device = new MmuDevice(
           buffer.getId(),
-          buffer.getWays(),
-          buffer.getSets(),
+          buffer.getWays().longValue(),
+          buffer.getSets().longValue(),
           address,
           addressFormat.getTagExpr(),
           addressFormat.getIndexExpr(),
